[WOM-108] Add thresholds to MccsFNDH and MccsSmartBox attributes list

+++ src/ska_low_mccs_pasd/pasd_bus/pasd_bus_device.py
@@ -330,7 +330,7 @@ class MccsPasdBus(SKABaseDevice[PasdBusComponentManager]):
                 4,
                 tango.AttrWriteType.READ_WRITE,
             ),
-            ("HumidityThresholds", (int,), 4, tango.AttrWriteType.READ_WRITE),
+            ("HumidityThresholds", (float,), 4, tango.AttrWriteType.READ_WRITE),
             (
                 "CommsGatewayTemperatureThresholds",
                 (float,),
